basic auth state management implementation
Implemented basic authorization (logged in or out) state management to change the state of the page based on user authorization.

Made a login button component that creates a modal with the ability to change the auth state to true. (or exit the modal)

Made a logout button component with a modal with the ability to change the auth state to false. (or cancel and close the modal)

Added an example state change to the main app to show how they would be implemented.

The actual ability to change the auth state will, in the future, be managed by actual login credentials on the back end, but for now it simply switches "true" or "false" based on the click of a login or logout button.