Create a global footer component
Created a global footer component and implemented it into the app.js file to be available throughout the site.

Includes a copyright section with logo, contact info, and address, a social link section (with dummy links that just lead to the homepage of the social site), and a site map with functional page links that are already implemented and dummy links to be changed or removed at a later date.

It's responsive and changes based on bootstrap breakpoints (currently only using the 768px medium breakpoint to change layout), and has a max width that fits with the max width of the main site section that already exists. It maxes out at 1000 pixels.

I has basic styling to look decently pretty at all sizes.