Email change form and account deletion changes
Created an email change form and made it work using firebase methods.

Added form validation, and submission validation. (can be used for login if we want to do that later)

Allowed for redirection to homepage upon deletion.

Changed the account deletion methods slightly from the last few commits to make it not use a hook.